+++ src/auth/index.ts
@@ -6,7 +6,7 @@ import { Ed25519Keypair } from "@mysten/sui/keypairs/ed25519";
 import { OAuth } from "./oauth";
 import { AkordApi } from "../api/akord-api";
 import { Env } from "../types/env";
-import { decode, DEFAULT_STORAGE, JWTClient } from "./jwt";
+import { decode, defaultStorage, JWTClient } from "./jwt";
 import { BadRequest } from "../errors/bad-request";
 import { decodeSuiPrivateKey } from "@mysten/sui/cryptography";
 import { retry } from "../api/api-client";
@@ -63,12 +63,17 @@ export class Auth {
     this.authProvider = options.authProvider;
     this.clientId = options.clientId;
     this.redirectUri = options.redirectUri;
-    this.storage = options.storage || DEFAULT_STORAGE;
+    this.storage = options.storage || defaultStorage();
     this.jwtClient = new JWTClient({ storage: this.storage, env: this.env });
   }
 
   public setEnv(env: Env) {
     this.env = env;
+    this.jwtClient = new JWTClient({ storage: this.storage, env: this.env });
+  }
+
+  public setStorage(storage: Storage) {
+    this.storage = storage;
     this.jwtClient = new JWTClient({ storage: this.storage, env: this.env });
   }