+++ src/crypto/user-encryption.ts
@@ -4,7 +4,7 @@ import Keystore from './storage/keystore';
 import { IncorrectEncryptionKey } from '../errors/incorrect-encryption-key';
 import { logger } from '../logger';
 import { X25519KeyPair } from './keypair';
-import { DEFAULT_STORAGE, JWTClient } from '../auth/jwt';
+import { defaultStorage, JWTClient } from '../auth/jwt';
 import { Env } from '../types';
 import { AESEncryptedPayload } from './types';
 import * as bip39 from 'bip39';
@@ -25,7 +25,7 @@ export class UserEncryption {
   constructor(config: { encPrivateKey?: string, encPrivateKeyBackup?: string, storage?: Storage, env?: Env } = {}) {
     this.encPrivateKey = config.encPrivateKey;
     this.encPrivateKeyBackup = config.encPrivateKeyBackup;
-    this.storage = config.storage || DEFAULT_STORAGE;
+    this.storage = config.storage || defaultStorage();
     this.userId = new JWTClient(config).getUserId();
     this.sessionKeyPath = `${this.userId}_${SESSION_KEY_PATH}`;
     this.encryptedPasswordKeyPath = `${this.userId}_${ENCRYPTED_PASSWORD_KEY_PATH}`;