+++ src/errors/error.ts
@@ -4,9 +4,15 @@ export class AkordError extends Error {
   statusCode: number;
   requestId: string;
 
-  constructor(message: string, error?: Error) {
+  constructor(message: string, error?: any) {
     super(message);
     this.requestId = (<any>error)?.response?.headers?.['request-id'];
-    logger.error(error);
+    if (error && error.response) {
+      logger.error(`${error.response.status}: ${error.response.statusText}`);
+      logger.debug(error.response.headers);
+      logger.debug(error.response.data);
+    } else {
+      logger.error(error);
+    }
   }
 }